pyproject.toml
[project]
name = "article-hub"
version = "0.1.0"
description = ""
authors = [
    {name = "borismen",email = "boris.tikhonov.21@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<=5.2)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-mongodb-backend (>=5.2.3,<6.0.0)",
    "celery[redis] (>=5.6.0,<6.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
